requirements.txt
aimet-torch==1.32.1.post1; sys_platform == "linux" and python_version == "3.10"

# midas_quantized
timm==1.0.3

# ffnet_78s_quantized
scikit-image>0.21.0,<0.25


# yolov8_seg
seaborn==0.11.0
thop==0.1.1.post2209072238
ultralytics==8.0.193